+++ mitra_models/src/properties/constants.rs
@@ -1,3 +1,4 @@
+pub const INSTANCE_RSA_SECRET_KEY: &str = "instance_rsa_secret_key";
 pub const FEDERATED_TIMELINE_RESTRICTED: &str = "federated_timeline_restricted";
 pub const ETHEREUM_CURRENT_BLOCK: &str = "ethereum_current_block";
 pub const ETHEREUM_TOKEN_WAITLIST_MAP: &str = "token_waitlist_map";